selector/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
readability-lxml==0.8.1
rank-bm25==0.2.2
lxml_html_clean==0.2.0
